backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
pytesseract==0.3.10
pdf2image==1.16.3
openai
httpx
pillow==10.1.0
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 